First version of a way to use an external optimizer. (#36)
Requires a command to be executed that handles EVERYTHING and reads in the initial
parameters and returns the optimized ones with their fitness set.

Add an example

Thanks to BXH fix a few D->d and || ->&&

+++ src/org/ogolem/adaptive/AdaptiveParameters.java
@@ -54,7 +54,7 @@ import org.ogolem.generic.ContinuousProblem;
  * cache is not thread-safe. Therefore in a multi-threaded environment, it is
  * absolutely required to generate a copy of this class for each thread.
  * @author Johannes Dieterich
- * @version 2020-04-29
+ * @version 2020-07-25
  */
 public class AdaptiveParameters extends ContinuousProblem<Double> {
 
@@ -78,7 +78,7 @@ public class AdaptiveParameters extends ContinuousProblem<Double> {
      * @param noOfParamsPerKey
      * @param paramsForMethod
      */
-    public AdaptiveParameters(final int numberOfParameters, final int id, final String[] keys,
+    public AdaptiveParameters(final int numberOfParameters, final long id, final String[] keys,
             final int[] noOfParamsPerKey, final String paramsForMethod){
         super();
         this.parameters = new double[numberOfParameters];
@@ -145,7 +145,7 @@ public class AdaptiveParameters extends ContinuousProblem<Double> {
      * @param formattedParams the formatted parameters as in a .param file.
      * @param id the id of this set
      */
-    public AdaptiveParameters(final String[] formattedParams, final int id){
+    public AdaptiveParameters(final String[] formattedParams, final long id){
         super();
         this.myID = id;
